chore: Remove unused home view
