+++ apps/web/public/images/web3-notification-bunny.png
