+++ apps/web/public/images/web3-notification-check.png
