+++ apps/web/src/components/ColoredWordHeading.tsx
