+++ apps/web/src/components/GradientLogoSvg.tsx
