+++ apps/web/src/components/SlideSvg.tsx
