+++ apps/web/src/views/AffiliatesProgram/components/Overview/HowItWork.tsx
@@ -1,6 +1,6 @@
 import { Flex, Heading, Text, Card, Box } from '@pancakeswap/uikit'
 import { useTranslation } from '@pancakeswap/localization'
-import GradientLogo from 'views/Home/components/GradientLogoSvg'
+import GradientLogo from 'components/GradientLogoSvg'
 
 const HowItWork = () => {
   const { t } = useTranslation()
