--- apps/web/src/views/Home/components/Banners/images/index.ts
@@ -1,19 +1,0 @@
-import farmV3MigrationBunny from './farmV3MigrationBunny.png'
-import farmV3MigrationBunnyFull from './farmV3MigrationBunnyFull.png'
-import web3NotificationBgMobile from './web3-notification-bg-mobile.png'
-import web3NotificationBg from './web3-notification-bg.png'
-import web3NotificationBubbleMobile from './web3-notification-bubble-mobile.png'
-import web3NotificationBubble from './web3-notification-bubble.png'
-import web3NotificationBunny from './web3-notification-bunny.png'
-import web3NotificationCheck from './web3-notification-check.png'
-
-export {
-  farmV3MigrationBunny,
-  farmV3MigrationBunnyFull,
-  web3NotificationBg,
-  web3NotificationBgMobile,
-  web3NotificationBubble,
-  web3NotificationBubbleMobile,
-  web3NotificationBunny,
-  web3NotificationCheck,
-}
