+++ apps/web/src/views/Migration/bCake.tsx
@@ -20,7 +20,7 @@ import { useRouter } from 'next/router'
 import React, { useEffect, useRef, useState } from 'react'
 import { usePollFarmsWithUserData } from 'state/farms/hooks'
 import bCakeMigrationImage from 'views/Farms/images/bCakeMigration.png'
-import { farmV3MigrationBunnyFull } from 'views/Home/components/Banners/images'
+import farmV3MigrationBunnyFull from '../../../public/images/farmV3MigrationBunnyFull.png'
 import MigrationSticky, { TEXT } from './components/MigrationSticky'
 import { MigrationProgressSteps, Step } from './components/ProgressSteps'
 import MigrationFarmStep from './components/bCake/MigrationFarmStep'
