+++ apps/web/src/views/Migration/v3.tsx
@@ -17,7 +17,8 @@ import { useRouter } from 'next/router'
 import React, { useEffect, useRef, useState } from 'react'
 import { usePollFarmsWithUserData } from 'state/farms/hooks'
 import LiquidityFormProvider from 'views/AddLiquidityV3/formViews/V3FormView/form/LiquidityFormProvider'
-import { farmV3MigrationBunny, farmV3MigrationBunnyFull } from 'views/Home/components/Banners/images'
+import farmV3MigrationBunny from '../../../public/images/farmV3MigrationBunny.png'
+import farmV3MigrationBunnyFull from '../../../public/images/farmV3MigrationBunnyFull.png'
 import MigrationSticky, { TEXT } from './components/MigrationSticky'
 import { MigrationProgressSteps, Step } from './components/ProgressSteps'
 import OldFarm from './components/v3/Step1'
