+++ apps/web/src/views/PancakeSquad/components/BunniesSection/index.tsx
@@ -2,9 +2,9 @@ import { Box, Button, Flex, Text } from '@pancakeswap/uikit'
 import { NextLinkFromReactRouter } from '@pancakeswap/widgets-internal'
 
 import { useTranslation } from '@pancakeswap/localization'
-import ColoredWordHeading from 'views/Home/components/ColoredWordHeading'
+import ColoredWordHeading from 'components/ColoredWordHeading'
 import { LandingBodyWrapper } from 'views/PancakeSquad/styles'
-import { SlideSvgDark, SlideSvgLight } from 'views/Home/components/SlideSvg'
+import { SlideSvgDark, SlideSvgLight } from 'components/SlideSvg'
 import useTheme from 'hooks/useTheme'
 import bunniesConfig from './config'
 import { StyledBunnySectionContainer, StyledTextContainer } from './styles'
