+++ packages/localization/src/config/translations.json
@@ -879,10 +879,6 @@
   "Your Liquidity": "Your Liquidity",
   "Remove liquidity to receive tokens back": "Remove liquidity to receive tokens back",
   "Total supply": "Total supply",
-  "Market cap": "Market cap",
-  "Current emissions": "Current emissions",
-  "$%marketCap%": "$%marketCap%",
-  "%cakeEmissions%/block": "%cakeEmissions%/block",
   "Harvest all": "Harvest all",
   "Copied": "Copied",
   "Make a Profile": "Make a Profile",
@@ -1570,7 +1566,6 @@
   "Received": "Received",
   "Claimable": "Claimable",
   "Remaining": "Remaining",
-  "Circulating Supply": "Circulating Supply",
   "Prizes will be announced and available for claiming at ~": "Prizes will be announced and available for claiming at ~",
   "Follow Update": "Follow Update",
   "Your Mines of Dalarnia - Bunny Helmet NFT will be airdropped to your wallet before 00:00 UTC on 2nd June.": "Your Mines of Dalarnia - Bunny Helmet NFT will be airdropped to your wallet before 00:00 UTC on 2nd June.",
@@ -2386,7 +2381,6 @@
   "Github": "Github",
   "Support": "Support",
   "Brand Assets": "Brand Assets",
-  "Token Burn": "Token Burn",
   "Ecosystem": "Ecosystem",
   "Unstaking from v2 farm?": "Unstaking from v2 farm?",
   "Migrate to v3 here": "Migrate to v3 here",
@@ -3708,7 +3702,6 @@
   "Weekly breakdown of CAKE emissions allocation by product": "Weekly breakdown of CAKE emissions allocation by product",
   "YTD Emissions": "YTD Emissions",
   "Year-to-date (YTD) Emissions: Absolute total (not the net amount) of CAKE emissions since start of the year": "Year-to-date (YTD) Emissions: Absolute total (not the net amount) of CAKE emissions since start of the year",
-  "Total Supply": "Total Supply",
   "Net Mint": "Net Mint",
   "Supply & Net Mint": "Supply & Net Mint",
   "Weekly net mint added to CAKE supply and the corresponding CAKE supply": "Weekly net mint added to CAKE supply and the corresponding CAKE supply",
